pyproject.toml
[project]
name = "tapeback"
version = "0.8.10"
description = "Local meeting recorder with transcription and speaker diarization for Obsidian"
readme = "README.md"
authors = [{ name = "Yast", email = "yastcher@gmail.com" }]
requires-python = ">=3.13"
license = { text = "Apache-2.0" }
keywords = ["meeting", "transcription", "whisper", "diarization", "obsidian", "pipewire"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "faster-whisper>=1.0.0",
    "pydantic-settings>=2.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
diarize = [
    "pyannote-audio>=3.1.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
llm = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
]
tray = [
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
]

[project.scripts]
tapeback = "tapeback.cli:cli"

[project.urls]
Homepage = "https://github.com/yastcher/tapeback"
Repository = "https://github.com/yastcher/tapeback"
Issues = "https://github.com/yastcher/tapeback/issues"
Changelog = "https://github.com/yastcher/tapeback/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tapeback"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pytest>=9.0.0",
    "pytest-cov>=6.0.0",
    "ty>=0.0.1a7",
    "pyannote-audio>=3.1.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = [
    ".claude",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "G", "PLC", "PLR", "SIM", "S", "RUF"]
ignore = [
    "S603",    # subprocess calls are safe (list-based, no shell)
    "S607",    # partial executable paths (ffmpeg, parecord) are intentional
    "PLR2004", # magic numbers — too noisy for audio/test code
    "S108",    # /tmp/tapeback is intentional, secured with 0700 permissions
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]  # assert, fake tokens in tests
"src/tapeback/cli.py" = ["PLC0415"]         # lazy imports for heavy ML deps
"src/tapeback/pipeline.py" = ["PLR0913"]  # pipeline functions take many params
"src/tapeback/channel.py" = ["PLR0912", "PLR0915"]  # complex audio analysis algorithms
"src/tapeback/diarizer.py" = ["PLC0415", "PLR0912", "PLR0915"]  # lazy ML imports, complex audio algorithms
"src/tapeback/summarizer.py" = ["PLC0415"]  # lazy imports for optional LLM SDKs

[tool.ruff.lint.isort]
known-local-folder = ["tapeback", "tests"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=tapeback --cov-report=term-missing"

[tool.coverage.run]
source = ["src/tapeback"]
omit = ["src/tapeback/tray.py"]  # GUI module, requires X display — tested locally only

[tool.coverage.report]
show_missing = true
fail_under = 85
exclude_lines = [
    "if TYPE_CHECKING:",
    "pragma: no cover",
]

[tool.ty.environment]
python-version = "3.13"
# Project root needed for `from tests.fixtures import ...`
# src/ needed for `from tapeback import ...`
extra-paths = [".", "src"]
